+++ src/main/java/insane96mcp/shieldsplus/setup/SPEnchantments.java
@@ -19,4 +19,5 @@ public class SPEnchantments {
     public static final RegistryObject<Enchantment> FAST_RECOVERY = ENCHANTMENTS.register("fast_recovery", FastRecoveryEnchantment::new);
     public static final RegistryObject<Enchantment> SHIELD_BASH = ENCHANTMENTS.register("shield_bash", ShieldBashEnchantment::new);
     public static final RegistryObject<Enchantment> PERFECT_PARRY = ENCHANTMENTS.register("perfect_parry", PerfectParryEnchantment::new);
+    public static final RegistryObject<Enchantment> CELESTIAL_GUARDIAN = ENCHANTMENTS.register("celestial_guardian", CelestialGuardianEnchantment::new);
 }