+++ src/main/java/insane96mcp/shieldsplus/world/item/enchantment/RecoilEnchantment.java
@@ -16,7 +16,7 @@ import net.minecraftforge.event.entity.living.ShieldBlockEvent;
 public class RecoilEnchantment extends Enchantment implements IBlockingEffect, IEnchantmentTooltip {
 
 	public RecoilEnchantment() {
-		super(Rarity.UNCOMMON, SPShieldItem.SHIELD, new EquipmentSlot[]{EquipmentSlot.MAINHAND, EquipmentSlot.OFFHAND});
+		super(Rarity.UNCOMMON, SPShieldItem.ENCHANTMENT_CATEGORY, new EquipmentSlot[]{EquipmentSlot.MAINHAND, EquipmentSlot.OFFHAND});
 	}
 
 	public int getMinCost(int p_44598_) {